first commit of two c files
both can compile and run, but still need to deal with edge cases such as gcd(a, b), a and/or b <0, a and/or b == 0, a is a multiple of b or b is a multiple of a, etc. Also, need to add general input validations, for instance, the number of inputs, a and/or b not integer, etc.